requirements.txt
flask==3.0.3
flask_sqlalchemy==3.1.1
pymysql==1.1.1
waitress==3.0.2
pandas==2.1.3
openpyxl==3.1.5
reportlab==4.0.8
weasyprint==61.2
# Melhorias de performance e segurança
flask-caching==2.1.0
flask-limiter==3.5.0
marshmallow==3.21.0
python-dateutil==2.8.2
# Testes
pytest==8.2.2
pytest-cov==4.1.0
pytest-flask==1.3.0
